+++ listerpoc/example/informers/externalversions/core/interface.go
@@ -3,7 +3,7 @@ package core
 import (
 	"k8s.io/client-go/informers/core"
 
-	v1 "github.com/kcp-dev/client-gen/listerpoc/listerpoc/example/informers/core/v1"
+	v1 "github.com/kcp-dev/client-gen/listerpoc/example/informers/externalversions/core/v1"
 )
 
 type Interface interface {
