+++ listerpoc/example/informers/externalversions/core/v1/configmaps.go
@@ -1,8 +1,10 @@
+package v1
+
 import (
 	corev1informers "k8s.io/client-go/informers/core/v1"
 	"k8s.io/client-go/tools/cache"
 
-	v1 "github.com/kcp-dev/client-gen/listerpoc/listerpoc/example/listers/core/v1"
+	v1 "github.com/kcp-dev/client-gen/listerpoc/example/listers/core/v1"
 )
 
 type ConfigMapInformer interface {
@@ -14,10 +16,10 @@ type configMapInformer struct {
 	delegate corev1informers.ConfigMapInformer
 }
 
-func (r realclusterConfigMapInformer) Informer() cache.SharedIndexInformer {
+func (r *configMapInformer) Informer() cache.SharedIndexInformer {
 	return r.delegate.Informer()
 }
 
-func (r realclusterConfigMapInformer) Lister() ConfigMapLister {
+func (r *configMapInformer) Lister() v1.ConfigMapLister {
 	return v1.NewConfigMapLister(r.delegate.Informer().GetIndexer())
 }
